requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.2
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
